requirements.txt
tabulate==0.8.7
pyyaml
pyjson
pdftotext
bs4
gensim==3.8.3
ipdb==0.13.4
matplotlib==3.3.3
nltk==3.5
numpy==1.19.4
pandas==1.1.5
Pattern==3.6
pattern3==3.0.0
pyLDAvis==2.1.2
regex==2020.11.13
requests==2.25.1
scikit-learn==0.23.2
scipy==1.5.4
simplejson==3.17.2
sklearn==0.0
spacy==2.3.5
tqdm==4.54.1
urllib3==1.26.2
webencodings==0.5.1
crossref-common
wordcloud
